Fix : " Map 수정 "
GameServer
 - Map 수정
  - _MapID 자료형 변경 int64 -> int16
  - MapInit() 수정
   - 맵 타일 정보 읽어올 때 DB에 기록되어 있는 타일 정보 읽어오기 추가
     맵 타일 정보 읽어온 후 읽어온 정보로 맵 타일 정보 초기화

+++ GameServer/Map.h
@@ -61,7 +61,7 @@ struct st_AStarNodeInt
 class CMap
 {
 public:
-	int64 _MapID;
+	int16 _MapID;
 
 	wstring _MapName;
 
@@ -97,7 +97,7 @@ public:
 	CMap();
 	~CMap();
 
-	void MapInit(int64 MapID, wstring MapName, int32 SectorSize, int8 ChannelCount);
+	void MapInit(int16 MapID, wstring MapName, int32 SectorSize, int8 ChannelCount);
 
 	CSector* GetSector(st_Vector2Int CellPosition);
 	CSector* GetSector(int32 IndexY, int32 IndexX);
